wrap bytes on bpf-loader-2 write instruction
