+++ explorer/src/components/instruction/RawDetails.tsx
@@ -1,15 +1,7 @@
 import React from "react";
 import { TransactionInstruction } from "@solana/web3.js";
 import { Address } from "components/common/Address";
-
-function wrap(input: string, length: number): string {
-  var result = [];
-  while (input.length) {
-    result.push(input.substr(0, length));
-    input = input.substr(length);
-  }
-  return result.join("\n");
-}
+import { wrap } from "utils";
 
 export function RawDetails({ ix }: { ix: TransactionInstruction }) {
   const data = wrap(ix.data.toString("hex"), 50);
